+++ CHANGELOG.md
@@ -1,10 +1,22 @@
 # Changelog
 
-All notable changes to `stdnum-php` will be documented in this file.
+All notable changes to `react-jitsi` will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.2]
+
+### ADDED
+
+- Added `<VirtualBackgroundSelector>` component to render the available effects
+- Added `VirtualBackgroundEffect` type
+
+### MODIFIED
+
+- `<JitsiMeeting>` and `<JitsiProvider>` now have the `virtualBackgroundEffects` property to allow the addition of default effects
+- `VirtualBackgroundType` no longer has the `none` option, to remove the effect, simply call the `removeVirtualBackground` or `setVirtualBackground` function with the null parameter.
+- Renamed `customEffect` to `effect` in `VirtualBackgroundConfig` type, and now is required property
 
 ## [0.3.1] - 2026-04-27